Task-Orchestrator-Service/requirements.txt
pika==1.3.2
redis==4.6.0
pydantic==2.0.2
pydantic-settings==2.0.1
PyYAML==6.0
